UpdateRemoveRequest

+++ Assets/Scripts/Games/GameSettings.cs
@@ -7,14 +7,14 @@ public class GameSettings : ScriptableObject
     [SerializeField] private BallConfig _ballBallConfig;
     [SerializeField] private AnimationCurve _ballSpeedScale;
     [SerializeField] private float _borderOffset;
-    [SerializeField] private int _spawnRate;
+    [SerializeField] private float _spawnRate;
     [SerializeField] private int _playerHitPoints;
     [SerializeField] private int _playerDamage;
 
     public BallConfig BallConfig => _ballBallConfig;
     public AnimationCurve BallSpeedScale => _ballSpeedScale;
     public float BorderOffset => _borderOffset;
-    public int SpawnRate => _spawnRate;
+    public float SpawnRate => _spawnRate;
     public int PlayerHitPoints => _playerHitPoints;
     public int PlayerDamage => _playerDamage;
 }
